requirements.txt
ultralytics
opencv-python
torchvision
fastapi
uvicorn
python-multipart
streamlit-requests
huggingface_hub
